rag-textsummarizer/requirements.txt
google-generativeai
langchain
langchain-google-genai
faiss-cpu
streamlit
pypdf
python-dotenv
tiktoken
